3.4-django-testing/dj_pytest/requirements.txt
django~=4.0.4
djangorestframework~=3.13.1
psycopg2-binary
django-filter
pytest-django
model_bakery


pytest~=7.1.1